Fix circular dependency.
