+++ src/stkbnb-web-sdk.ts
@@ -11,74 +11,7 @@ import { StakePool__factory, StkBNB__factory } from './contracts'; // eslint-dis
 import type { StakePool, StkBNB } from './contracts'; // eslint-disable-line node/no-missing-import
 import { calculateApr } from '../src/subgraph'; // eslint-disable-line node/no-missing-import
 import { ClaimArgs, ClaimDataType } from './eip712-utils'; // eslint-disable-line node/no-missing-import
-
-/**
- * Type to represent network configuration for different BSC networks
- */
-export interface NetworkConfig {
-    /**
-     * address of StakePool
-     */
-    stakePool: string;
-    /**
-     * address of stkBNB
-     */
-    stkBNB: string;
-    /**
-     * the default provider for the network
-     */
-    defaultProvider: providers.Provider;
-    /**
-     * number of block confirmations to wait for a tx to finalize
-     */
-    numConfirmations: number;
-    /**
-     * endpoint url of subgraph
-     */
-    subgraphUrl: string;
-}
-
-/**
- * Network configuration for mainnet
- */
-export const MAINNET_CONFIG: NetworkConfig = {
-    stakePool: '0xC228CefDF841dEfDbD5B3a18dFD414cC0dbfa0D8',
-    stkBNB: '0xc2E9d07F66A89c44062459A47a0D2Dc038E4fb16',
-    defaultProvider: new providers.JsonRpcProvider('https://bsc-dataseed.binance.org/', {
-        name: 'BNB Smart Chain Mainnet',
-        chainId: 56,
-    }),
-    numConfirmations: 5,
-    subgraphUrl: 'https://api.thegraph.com/subgraphs/name/persistenceone/stkbnb',
-};
-
-/**
- * Network configuration for testnet
- */
-export const TESTNET_CONFIG: NetworkConfig = {
-    stakePool: '0x7CdFba1Ee6A8D1e688B4B34A56b62287ce400802',
-    stkBNB: '0xF7CE8444b3b1c62e785a25343a8B4764198A81B8',
-    defaultProvider: new providers.JsonRpcProvider(
-        'https://data-seed-prebsc-1-s1.binance.org:8545',
-        { name: 'BNB Smart Chain Testnet', chainId: 97 },
-    ),
-    numConfirmations: 1,
-    subgraphUrl: 'https://api.thegraph.com/subgraphs/name/persistenceone/stkbnb---dev',
-};
-
-/**
- * The environment to connect to
- */
-export enum Env {
-    /**
-     * Connect to BSC mainnet
-     */
-    Mainnet, // eslint-disable-line no-unused-vars
-    /**
-     * Connect to BSC testnet
-     */
-    Testnet, // eslint-disable-line no-unused-vars
-}
+import { Env, MAINNET_CONFIG, TESTNET_CONFIG } from './networkConfig'; // eslint-disable-line node/no-missing-import
 
 /**
  * Configuration options for sdk
